Clean up Function Evaluation
This required moving MemFS into kpt module to avoid circular dependency.
Also simplify the kpt module paths; eventually we need to rename it.
