--- porch/engine/pkg/engine/fs_test.go
@@ -1,11 +1,0 @@
-package engine
-
-import "testing"
-
-func TestCreate(t *testing.T) {
-	fs := &memfs{}
-
-	fs.MkdirAll("a/b/c/")
-	fs.MkdirAll("/d/e/f")
-	fs.WriteFile("/a/b/c/foo.yaml", []byte("Hello World"))
-}
