+++ porch/kpt/fs_test.go
@@ -1,0 +1,36 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package kpt
+
+import "testing"
+
+func TestMemFS(t *testing.T) {
+	fs := &memfs{}
+
+	if err := fs.MkdirAll("a/b/c/"); err != nil {
+		t.Errorf("MkdirAll(\"a/b/c/\") failed %v", err)
+	}
+	if err := fs.MkdirAll("/d/e/f"); err != nil {
+		t.Errorf("MkdirAll(\"/d/e/f\") failed: %v", err)
+	}
+	if err := fs.WriteFile("/a/b/c/foo.txt", []byte("Hello World")); err != nil {
+		t.Errorf("Failed to write file: %v", err)
+	}
+	if res, err := fs.ReadFile("/a/b/c/foo.txt"); err != nil {
+		t.Errorf("Failed to read file: %v", err)
+	} else if got, want := string(res), "Hello World"; got != want {
+		t.Errorf("unexpected file contents: got %q, want %q", got, want)
+	}
+}
